remove backburner link from header
The link being in the header caused the entire section to become a link, so you couldn't even copy the text inside it. It also partially duplicated the header. I suspect this is some post-processing thing we do that making an error here (the markdown directly would probably be fine), but I don't think that's worth investigating/debugging - the link being below is good enough.

Moving the link out of the header should fix the issue :)