Updates the explanation of template helpers (#293)
* fix all uses of non-relative guides urls

* write test to check for non-relative urls

* Simplifies the "All we want to do..." portion

On the Routes & Templates page, in the An Index Route section, the
paragraphs starting at "All we want to do..." presents too much
ancillary data. This change simplifies it to focus on just the
beforeModel() hook and the replaceWith() function.

* Update guides/v3.5.0/tutorial/routes-and-templates.md

Co-Authored-By: zachgarwood <iam@zachgarwood.com>

* Clarifies where template properties come from

On the Building a Simple Component page, we should explicitly call out
that properties of components are available to their templates. This
change does so.

This addresses one of the "tiny tutorial fixes" @jenweber listed
in #245.

* Updates the explanation of template helpers

On the Creating a Handlebars Helper page, we understate the purpose and
usage template helpers. This change modifies the opening paragraphs to
go into more details.

* Update guides/v2.10.0/tutorial/hbs-helper.md

Co-Authored-By: zachgarwood <iam@zachgarwood.com>

* Update guides/v2.10.0/tutorial/hbs-helper.md

Co-Authored-By: zachgarwood <iam@zachgarwood.com>

* Removed duplicate phrase

* Fixed url (#295)

* Fixed url

The absolute url was breaking [this test](https://travis-ci.org/ember-learn/guides-source/builds/461958652#L3463). I made it relative.

* Nest link one level deeper

* Adds Handlebars explainer paragraph (#291)

* Adds Handlebars explainer paragraph

In Building a Simple Component, we mention Handlebars templates but
never state what Handlebars. This change adds a brief paragraph to
describe Handlebars, link to a more robust explanation (Handlebars
Basics), and explain the relationship between templates and component
source files.

* remove comma